+++ Modules/Modules.AQEDecimal/AQCI/AQCIResult.cs
@@ -8,7 +8,7 @@
         /// <summary>
         /// 空气质量综合指数
         /// </summary>
-        public double? AQCI { get; set; }
+        public decimal? AQCI { get; set; }
         /// <summary>
         /// 首要污染物
         /// </summary>
